Avoid crash when deduce fails for imported generic
An imported generic has bindings which are of type ImportRefLoaded, and
if they come from another package, they have no entity name attached to
them.

To get the entity name, we need to get the equivalent instruction for
the binding from the original SemIR::File. We then have to convert to a
string explicitly instead of passing a `NameId` to the diagnostic
emitter, since it will only look up IDs in the local SemIR::File.