Make the tests_file.tests manifest path relative to the binary
Neither rootpath nor execpath work for all cases that `file_test` is run
under.

The rootpath makes a path relative to output directory, such as
`toolchain/testing/file_test.tests`. When run in a debugger, the CWD is
at the repository root, so this path does not point to the file
correctly. It needs to include the root of the build directory.

The execpath makes a path relative to the repository root, such as
`bazel-bin/toolchain/testing/file_test.tests`. This works in a debugger
from the repository root (though not from arbitrary places). But it
fails under `bazel run`, which runs `file_test` with a CWD in the output
directory. https://stackoverflow.com/a/70265855 points out this is by
design and there's no good way to change the CWD for bazel run.

So this change makes the manifest file path relative to the binary
instead of either rootpath or execpath. The `file_test` binary takes the
path to the exe, moves to the parent path (drops the binary itself from
the path), and appends the manifest path specified from bazel. This
produces an absolute path to the tests_file.tests that is always correct.

This change attempts to make no change to explorer, assuming it is
working as-is.