Perform member lookup on FacetAccessType
The name scope lookup and member name lookup both need to handle the
case where the base inst is a FacetAccessType. Then we move from the
FacetAccessType to the FacetType which is the type of the instruction
inside the FacetAccessType.

We do name lookup on FacetType already, so "unwrapping" the
FacetAccessType to the FacetType just makes use of that path.
Similarly we do member access on FacetType already, so we can reuse that
codepath with the FacetType found in the FacetAccessType.